+++ root/config.php
@@ -39,6 +39,9 @@ define('MAX_WIDTH', 720);
 // Maximum number of statuses allowed in each feed
 define('MAX_STATUSES', 50);
 
+// Number of statuses generated each time a queue job runs
+define('STATUS_JOB_BATCH_SIZE', 3);
+
 // Maximum days to keep images. Default is 360 and it should be kept at
 // least this value in production.
 define('IMG_AGE', 360);